requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
brapi==1.0.1
certifi==2025.10.5
charset-normalizer==3.4.4
distro==1.9.0
Django==4.1.13
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
requests==2.32.5
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
pika==1.3.2        
psycopg[binary]>=3.2.1
requests==2.32.5
python-dotenv==1.2.1
python-telegram-bot[all]>=21.0